Fix resource loading
Most importantly: sizeof a data pointer is assumed to be 4, and fit into
int. Use intptr_t where an integer holding a pointer is expected. Use
pointer sizes in allocations. Somehow, nothing seems to break because of
gaps in structure arrays and such.

Enable Werr, but disable some warnings.

A DOOM1 wad seems to be loading fine, but Xlib stuff is still failing
(expecting to create a 8-bit depth PseudoColor screen).

+++ linuxdoom-1.10/doomstat.h
@@ -36,6 +36,8 @@
 // We need the playr data structure as well.
 #include "d_player.h"
 
+#include <stdint.h>
+
 
 #ifdef __GNUG__
 #pragma interface
@@ -106,8 +108,8 @@ extern  boolean	deathmatch;
 //  Sound FX volume has default, 0 - 15
 //  Music volume has default, 0 - 15
 // These are multiplied by 8.
-extern int snd_SfxVolume;      // maximum volume for sound
-extern int snd_MusicVolume;    // maximum volume for music
+extern intptr_t snd_SfxVolume;      // maximum volume for sound
+extern intptr_t snd_MusicVolume;    // maximum volume for music
 
 // Current music/sfx card - index useless
 //  w/o a reference LUT in a sound module.
@@ -252,7 +254,7 @@ extern  boolean         precache;
 //  to force a wipe on the next draw
 extern  gamestate_t     wipegamestate;
 
-extern  int             mouseSensitivity;
+extern  intptr_t        mouseSensitivity;
 //?
 // debug flag to cancel adaptiveness
 extern  boolean         singletics;
